Fix for cypress tests not starting
